requirements.txt
# Database & Connection
psycopg2-binary
asyncpg
redis
pgvector
langsmith

# Data Processing
pandas
openpyxl
numpy

# LangChain - Compatible versions
langchain
langchain-core
langchain-community
langchain-openai
langgraph
langgraph-checkpoint-redis

# Pydantic (important for compatibility)
pydantic
pydantic-core

# Vector Embeddings & AI
FlagEmbedding
sentence-transformers
transformers==4.44.2
peft==0.12.0
torch

# Environment
python-dotenv

